fix(segmentation): segmentation service listeners should be added in onModeEnter and not in preRegistration. This permits the listeners to be executed on a subsequent navigation into a mode.
