+++ extensions/cornerstone/src/init.tsx
@@ -39,10 +39,6 @@ import { useLutPresentationStore } from './stores/useLutPresentationStore';
 import { usePositionPresentationStore } from './stores/usePositionPresentationStore';
 import { useSegmentationPresentationStore } from './stores/useSegmentationPresentationStore';
 import { imageRetrieveMetadataProvider } from '@cornerstonejs/core/utilities';
-import {
-  setupSegmentationDataModifiedHandler,
-  setupSegmentationModifiedHandler,
-} from './utils/segmentationHandlers';
 
 const { registerColormap } = csUtilities.colormap;
 
@@ -194,18 +190,6 @@ export default async function init({
     });
   });
 
-  // Setup segmentation event handlers
-  const { unsubscribe: unsubscribeSegmentationDataModifiedHandler } =
-    setupSegmentationDataModifiedHandler({
-      segmentationService,
-      customizationService,
-      commandsManager,
-    });
-
-  const { unsubscribe: unsubscribeSegmentationModifiedHandler } = setupSegmentationModifiedHandler({
-    segmentationService,
-  });
-
   // When a custom image load is performed, update the relevant viewports
   hangingProtocolService.subscribe(
     hangingProtocolService.EVENTS.CUSTOM_IMAGE_LOAD_PERFORMED,
@@ -312,47 +296,6 @@ export default async function init({
 
   // Call this function when initializing
   initializeWebWorkerProgressHandler(servicesManager.services.uiNotificationService);
-
-  const { unsubscribe: unsubscribeSegmentationCreated } = segmentationService.subscribe(
-    segmentationService.EVENTS.SEGMENTATION_ADDED,
-    evt => {
-      const { segmentationId } = evt;
-      const displaySet = displaySetService.getDisplaySetByUID(segmentationId);
-      if (displaySet) {
-        return;
-      }
-
-      const segmentation = segmentationService.getSegmentation(segmentationId);
-      const label = segmentation.cachedStats.info;
-      const imageIds = segmentation.representationData.Labelmap.imageIds;
-
-      // Create a display set for the segmentation
-      const segmentationDisplaySet = {
-        displaySetInstanceUID: segmentationId,
-        SOPClassUID: '1.2.840.10008.5.1.4.1.1.66.4',
-        SOPClassHandlerId: '@ohif/extension-cornerstone-dicom-seg.sopClassHandlerModule.dicom-seg',
-        SeriesDescription: label,
-        Modality: 'SEG',
-        numImageFrames: imageIds.length,
-        imageIds,
-        isOverlayDisplaySet: true,
-        label,
-        madeInClient: true,
-        segmentationId: segmentationId,
-        isDerived: true,
-      };
-
-      displaySetService.addDisplaySets(segmentationDisplaySet);
-    }
-  );
-
-  const unsubscriptions = [
-    unsubscribeSegmentationDataModifiedHandler,
-    unsubscribeSegmentationModifiedHandler,
-    unsubscribeSegmentationCreated,
-  ];
-
-  return { unsubscriptions };
 }
 
 function initializeWebWorkerProgressHandler(uiNotificationService) {
